% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amberdata.R
\name{amberdata_api_call}
\alias{amberdata_api_call}
\title{amberdata_api_call}
\usage{
amberdata_api_call(url, api_key, method, query = NULL)
}
\arguments{
\item{url}{the url for your Amberdata API call}

\item{api_key}{your Amberdata API key}

\item{method}{"GET" or "POST"}

\item{query}{your query parameters. The default value is NULL.}
}
\value{
returns data from your Amberdata API call
}
\description{
amberdata_api_call
}
\examples{
\dontrun{
api_key <- "..."
url <- "https://web3api.io/api/v2/market/exchanges"
method <- "GET"
exchanges <- amberdata_api_call(url, api_key, method)}
}
