% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_us.R
\name{binance_us_signature}
\alias{binance_us_signature}
\title{binance_us_signature}
\usage{
binance_us_signature(data, secret)
}
\arguments{
\item{data}{your URL encoded query parameters}

\item{secret}{your Binance.US secret key}
}
\value{
returns your Binance.US signature for use in API calls
}
\description{
binance_us_signature
}
\examples{
\dontrun{
time <- binance_us_time()
data <- paste('timestamp=', time, sep = '')
secret <- "..."
signature <- binance_us_signature(data, secret)}
}
