% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_us.R
\name{binance_us_time}
\alias{binance_us_time}
\title{binance_us_time}
\usage{
binance_us_time()
}
\value{
returns a timestamp in the format that Binance.US expects
}
\description{
binance_us_time
}
\examples{
binance_us_time()
}
