% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain_dot_com.R
\name{blockchain_dot_com_l3_order_book}
\alias{blockchain_dot_com_l3_order_book}
\title{blockchain_dot_com_l3_order_book}
\usage{
blockchain_dot_com_l3_order_book(symbol)
}
\arguments{
\item{symbol}{the symbol for which to retrieve data}
}
\value{
returns a list containing a 'bids' and an 'asks' dataframe along with
the specified symbol
}
\description{
blockchain_dot_com_l3_order_book
}
\examples{
\dontrun{
symbol <- 'BTC-USD'
l3_order_book <- blockchain_dot_com_l3_order_book(symbol)
bids <- l3_order_book$bids
asks <- l3_order_book$asks}
}
