% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain_dot_com.R
\name{blockchain_dot_com_symbol}
\alias{blockchain_dot_com_symbol}
\title{blockchain_dot_com_symbol}
\usage{
blockchain_dot_com_symbol(symbol)
}
\value{
returns a list with various data for specified symbol
}
\description{
blockchain_dot_com_symbol
}
\examples{
\dontrun{
symbol <- 'BTC-USD'
symbol_data <- blockchain_dot_com_symbol(symbol)}
}
