% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain_dot_com.R
\name{blockchain_dot_com_symbols}
\alias{blockchain_dot_com_symbols}
\title{blockchain_dot_com_symbols}
\usage{
blockchain_dot_com_symbols(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a list with various data for all symbols
}
\description{
blockchain_dot_com_symbols
}
\examples{
blockchain_dot_com_symbols(4.5)
}
