% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain_dot_com.R
\name{blockchain_dot_com_tickers}
\alias{blockchain_dot_com_tickers}
\title{blockchain_dot_com_tickers}
\usage{
blockchain_dot_com_tickers(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe with price and volume data for all symbols
}
\description{
blockchain_dot_com_tickers
}
\examples{
blockchain_dot_com_tickers(4.5)
}
