% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cex_io.R
\name{cex_io_balance}
\alias{cex_io_balance}
\title{cex_io_balance}
\usage{
cex_io_balance(username, api_key, api_secret, timeout_seconds = 60)
}
\arguments{
\item{username}{your cex.io username}

\item{api_key}{your cex.io api_key}

\item{api_secret}{your cex.io api_secret}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a list with your balances for each currency
}
\description{
cex_io_balance
}
\examples{
\dontrun{
username <- "..."
api_key <- "..."
api_secret <- "..."
balances <- cex_io_balance(username, api_key, api_secret, 4.5)}
}
