% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cex_io.R
\name{cex_io_converter}
\alias{cex_io_converter}
\title{cex_io_converter}
\usage{
cex_io_converter(symbol_1, symbol_2, amount)
}
\arguments{
\item{symbol_1}{the first currency in your pair}

\item{symbol_2}{the second currency in your pair}

\item{amount}{the currency amount to convert denominated in symbol_1}
}
\value{
returns the converted amount denominated in symbol_2
}
\description{
cex_io_converter
}
\examples{
\dontrun{
symbol_1 <- 'btc'
symbol_2 <- 'usd'
amount <- '2.5'
converted <- cex_io_converter(symbol_1, symbol_2, amount)}
}
