% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cex_io.R
\name{cex_io_currency_limits}
\alias{cex_io_currency_limits}
\title{cex_io_currency_limits}
\usage{
cex_io_currency_limits()
}
\value{
returns a dataframe with information about currency limits on CEX.io.
}
\description{
cex_io_currency_limits
}
\examples{
\dontrun{
currency_limits <- cex_io_currency_limits()}
}
