% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cex_io.R
\name{cex_io_nonce}
\alias{cex_io_nonce}
\title{cex_io_nonce}
\usage{
cex_io_nonce()
}
\value{
returns a nonce for use in your signature
}
\description{
cex_io_nonce
}
\examples{
\dontrun{
nonce <- cex_io_nonce()}
}
