% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cex_io.R
\name{cex_io_signature}
\alias{cex_io_signature}
\title{cex_io_signature}
\usage{
cex_io_signature(username, api_key, api_secret, nonce)
}
\arguments{
\item{username}{your cex.io username}

\item{api_key}{your cex.io api_key}

\item{api_secret}{your cex.io api_secret}

\item{nonce}{a nonce to use in your signature and request body}
}
\value{
returns a signature for use in your API call
}
\description{
cex_io_signature
}
\examples{
\dontrun{
nonce <- cex_io_nonce()
username <- "..."
api_key <- "..."
api_secret <- "..."
sig <- cex_io_signature(username, api_key, api_secret, nonce)}
}
