% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cex_io.R
\name{cex_io_ticker}
\alias{cex_io_ticker}
\title{cex_io_ticker}
\usage{
cex_io_ticker(symbol_1, symbol_2)
}
\arguments{
\item{symbol_1}{the first currency in your pair}

\item{symbol_2}{the second currency in your pair}
}
\value{
returns a list with basic trading information about your specified currency pair for the last 24 hours.
}
\description{
cex_io_ticker
}
\examples{
\dontrun{
symbol_1 <- 'btc'
symbol_2 <- 'usd'
ticker_info <- cex_io_ticker(symbol_1, symbol_2)}
}
