% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase.R
\name{coinbase_accounts}
\alias{coinbase_accounts}
\title{coinbase_accounts}
\usage{
coinbase_accounts(api_key, api_secret)
}
\arguments{
\item{api_key}{your Coinbase API key}

\item{api_secret}{your Coinbase API secret}
}
\value{
returns a dataframe with information about your Coinbase accounts
}
\description{
coinbase_accounts
}
\examples{
\dontrun{
api_key <- "..."
api_secret <- "..."
accounts <- coinbase_accounts(api_key, api_secret)}
}
