% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase.R
\name{coinbase_all_currencies}
\alias{coinbase_all_currencies}
\title{coinbase_all_currencies}
\usage{
coinbase_all_currencies(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe with information about all currencies known to Coinbase
}
\description{
coinbase_all_currencies
}
\examples{
coinbase_all_currencies(4.5)
}
