% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase.R
\name{coinbase_signature}
\alias{coinbase_signature}
\title{coinbase_signature}
\usage{
coinbase_signature(api_secret, coinbase_time, method, path, body)
}
\arguments{
\item{api_secret}{your Coinbase API secret}

\item{coinbase_time}{a timestamp in the correct format according to Coinbase}

\item{method}{"GET" or "POST"}

\item{path}{the path of your API call}

\item{body}{the body of your API call}
}
\value{
returns a signature for use in your Coinbase API calls
}
\description{
coinbase_signature
}
\examples{
\dontrun{
api_secret <- "..."
coinbase_time <- coinbase_time()
method <- "GET"
path <- "/api/v3/brokerage/accounts"
body <- ""
coinbase_signature <- coinbase_signature(api_secret, coinbase_time, method, path, body)}
}
