% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase.R
\name{coinbase_single_currency}
\alias{coinbase_single_currency}
\title{coinbase_single_currency}
\usage{
coinbase_single_currency(currency)
}
\arguments{
\item{currency}{the currency id for the relevant asset}
}
\value{
returns a list with details related to the specified currency
}
\description{
coinbase_single_currency
}
\examples{
currency <- 'btc'
coinbase_single_currency(currency)
}
