% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase.R
\name{coinbase_time}
\alias{coinbase_time}
\title{coinbase_time}
\usage{
coinbase_time()
}
\value{
returns a timestamp for use in your Coinbase API calls
}
\description{
coinbase_time
}
\examples{
\dontrun{
coinbase_time <- coinbase_time()}
}
