% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko.R
\name{coingecko_categories}
\alias{coingecko_categories}
\title{coingecko_categories}
\usage{
coingecko_categories(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe of all categories on CoinGecko.
}
\description{
coingecko_categories
}
\examples{
coingecko_categories(4.5)
}
