% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko.R
\name{coingecko_global_data}
\alias{coingecko_global_data}
\title{coingecko_global_data}
\usage{
coingecko_global_data(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a list containing high-level statistics about the
cryptocurrency ecosystem.
}
\description{
coingecko_global_data
}
\examples{
coingecko_global_data(4.5)
}
