% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko.R
\name{coingecko_ping}
\alias{coingecko_ping}
\title{coingecko_ping}
\usage{
coingecko_ping()
}
\value{
returns the Coingecko API server status
}
\description{
coingecko_ping
}
\examples{
\dontrun{
status <- coingecko_ping()}
}
