% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko.R
\name{coingecko_price}
\alias{coingecko_price}
\title{coingecko_price}
\usage{
coingecko_price(
  id,
  vs_currency,
  include_market_cap = NULL,
  include_24hr_vol = NULL,
  include_24hr_change = NULL,
  include_last_updated_at = NULL,
  precision = NULL
)
}
\arguments{
\item{id}{one or more comma-separated asset ids to query}

\item{vs_currency}{one or more comma-separated vs_currencies to query}

\item{include_market_cap}{optionally provide a 'true' or 'false' value to
include/exclude market cap. The default is 'false'.}

\item{include_24hr_vol}{optionally provide a 'true' or 'false' value to
include/exclude 24-hour volume. The default is 'false'.}

\item{include_24hr_change}{optionally provide a 'true' or 'false' value to
include/exclude the 24-hour price change. The default is 'false'.}

\item{include_last_updated_at}{optionally provide a 'true' or 'false' value to
include/exclude the last updated information. The default is 'false'.}

\item{precision}{optionally specify the decimal precision to return. Choose
either 'full' or any number between 0 and 18.}
}
\value{
returns a list of currency prices
}
\description{
coingecko_price
}
\examples{
\dontrun{
prices <- coingecko_price(id = 'bitcoin', vs_currency = 'usd')}
}
