% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinlist.R
\name{coinlist_fees}
\alias{coinlist_fees}
\title{coinlist_fees}
\usage{
coinlist_fees(api_key, api_secret)
}
\arguments{
\item{api_key}{your Coinlist API key}

\item{api_secret}{your Coinlist API secret}
}
\value{
returns a list containing Coinlist fees by symbols.
}
\description{
coinlist_fees
}
\examples{
\dontrun{
api_key <- "..."
api_secret <- "..."
fees <- coinlist_fees(api_key, api_secret)}
}
