% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinlist.R
\name{coinlist_signature}
\alias{coinlist_signature}
\title{coinlist_signature}
\usage{
coinlist_signature(api_secret, coinlist_time, method, path, body)
}
\arguments{
\item{api_secret}{your Coinlist API secret}

\item{coinlist_time}{a timestamp in the correct format according to Coinlist}

\item{method}{"GET" or "POST"}

\item{path}{the path of your API call}

\item{body}{the body of your API call}
}
\value{
returns a signature for use in your Coinlist API calls
}
\description{
coinlist_signature
}
\examples{
\dontrun{
api_secret <- "..."
coinlist_time <- coinlist_time()
method <- "GET"
path <- "/v1/accounts"
body <- ""
coinlist_signature <- coinlist_signature(api_secret, coinlist_time, method, path, body)}
}
