% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinlist.R
\name{coinlist_symbols}
\alias{coinlist_symbols}
\title{coinlist_symbols}
\usage{
coinlist_symbols()
}
\value{
returns a dataframe with information about symbols available on
Coinlist Pro
}
\description{
coinlist_symbols
}
\examples{
coinlist_symbols()
}
