% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinlist.R
\name{coinlist_time}
\alias{coinlist_time}
\title{coinlist_time}
\usage{
coinlist_time()
}
\value{
returns a timestamp for use in your Coinlist API calls
}
\description{
coinlist_time
}
\examples{
coinlist_time()
}
