% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmarketcap.R
\name{coinmarketcap_categories}
\alias{coinmarketcap_categories}
\title{coinmarketcap_categories}
\usage{
coinmarketcap_categories(
  api_key,
  start = "1",
  limit = NULL,
  id = NULL,
  slug = NULL,
  symbol = NULL
)
}
\arguments{
\item{api_key}{your CoinMarketCap API key}

\item{start}{you can use this parameter to offset your first result. The
default value is "1".}

\item{limit}{an optional string value between 1 and 5000 which tells
CoinMarketCap how many results to return. The default value is NULL.}

\item{id}{filter categories by one or more asset ids. The default value is
NULL. Multiple values must be comma-separated.}

\item{slug}{filter categories by one or more asset slugs. The default value is
NULL. Multiple values must be comma-separated.}

\item{symbol}{filter categories by one or more asset symbols. The default
value is NULL. Multiple values must be comma-separated.}
}
\value{
returns a datafrane with information about CoinMarketCap asset
categories.
}
\description{
coinmarketcap_categories
}
\examples{
\dontrun{
api_key <- "..."
categories <- coinmarketcap_categories(api_key)}
}
