% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covalent.R
\name{covalent_api_call}
\alias{covalent_api_call}
\title{covalent_api_call}
\usage{
covalent_api_call(url, method, query = NULL, csv = FALSE)
}
\arguments{
\item{url}{the Covalent URL for use in your API call}

\item{method}{'GET' or 'POST'}

\item{query}{your query parameters formatted as a named list}

\item{csv}{'TRUE' will return csv data parsed as a dataframe while 'FALSE'
will return json data. The default value is 'FALSE'.}
}
\value{
returns your Covalent API data
}
\description{
covalent_api_call
}
\examples{
\dontrun{
url <- "https://api.covalenthq.com/v1/1/address/trevorfrench.eth/balances_v2/"
api_key <- "..."
query <- list(key = api_key, format = NULL)
method <- "GET"
balances <- covalent_api_call(url, method, api_key, method, query, csv = FALSE)}
}
