% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covalent.R
\name{covalent_balances}
\alias{covalent_balances}
\title{covalent_balances}
\usage{
covalent_balances(
  api_key,
  chain_id,
  address,
  csv = FALSE,
  timeout_seconds = 60
)
}
\arguments{
\item{api_key}{your Covalent API key}

\item{chain_id}{the string id of the chain for which you wish to check
balances.}

\item{address}{the address you for which wish to check balances.}

\item{csv}{'TRUE' will return csv data parsed as a dataframe while 'FALSE'
will return json data. The default value is 'FALSE'.}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns either a list or a dataframe with account balances
}
\description{
covalent_balances
}
\examples{
\dontrun{
api_key <- "..."
balances <- covalent_balances(api_key, "1", "trevorfrench.eth", csv = FALSE)}
}
