% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_dot_com.R
\name{crypto_dot_com_get_trades}
\alias{crypto_dot_com_get_trades}
\title{crypto_dot_com_get_trades}
\usage{
crypto_dot_com_get_trades(instrument)
}
\arguments{
\item{instrument}{the instrument name which you want to query}
}
\value{
returns trade data for specified instrument. Refer to Crypto.com for
help interpreting response data:
https://exchange-docs.crypto.com/spot/index.html#public-get-trades
}
\description{
crypto_dot_com_get_trades
}
\examples{
\dontrun{
trades <- crypto_dot_com_get_trades("BTC_USDT")}
}
