% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_dot_com.R
\name{crypto_dot_com_instruments}
\alias{crypto_dot_com_instruments}
\title{crypto_dot_com_instruments}
\usage{
crypto_dot_com_instruments()
}
\value{
returns a dataframe with information about instruments available on Crypto.com
}
\description{
crypto_dot_com_instruments
}
\examples{
\dontrun{
instruments <- crypto_dot_com_instruments()}
}
