% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etherscan.R
\name{etherscan_account_balance}
\alias{etherscan_account_balance}
\title{etherscan_account_balance}
\usage{
etherscan_account_balance(address, api_key, tag = "latest")
}
\arguments{
\item{address}{the address for which you wish to retrieve the balance.}

\item{api_key}{your Etherscan API key}

\item{tag}{pre-defined block parameter, either earliest, pending or latest.
Default is latest.}
}
\value{
returns the balance for the specified address
}
\description{
etherscan_account_balance
}
\examples{
\dontrun{
address <- "0xde0b295669a9fd93d5f28d9ec85e40f4cb697bae"
api_key <- "..."
account_balance <- etherscan_account_balance(address, api_key)}
}
