% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etherscan.R
\name{etherscan_block_reward}
\alias{etherscan_block_reward}
\title{etherscan_block_reward}
\usage{
etherscan_block_reward(block, api_key)
}
\arguments{
\item{block}{the numeric block number}

\item{api_key}{your Etherscan API key}
}
\value{
returns the block and uncle reward for the specified block number as
a list.
}
\description{
etherscan_block_reward
}
\examples{
\dontrun{
block <- 12697906
api_key <- "..."
block_reward <- etherscan_block_reward(block, api_key)}
}
