% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etherscan.R
\name{etherscan_contract_abi}
\alias{etherscan_contract_abi}
\title{etherscan_contract_abi}
\usage{
etherscan_contract_abi(address, api_key, timeout_seconds = 60)
}
\arguments{
\item{address}{the contract address for which you wish to retrieve the ABI.}

\item{api_key}{your Etherscan API key}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns the contract ABI for the specified address
}
\description{
etherscan_contract_abi
}
\examples{
\dontrun{
address <- "0xfb6916095ca1df60bb79ce92ce3ea74c37c5d359"
api_key <- "..."
abi <- etherscan_contract_abi(address, api_key)}
}
