% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gemini_price_feed}
\alias{gemini_price_feed}
\title{gemini_price_feed}
\usage{
gemini_price_feed(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe containing pairs, their current price, and their
24 hour change in price
}
\description{
gemini_price_feed
}
\examples{
gemini_price_feed(4.5)
}
