% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gemini_symbols}
\alias{gemini_symbols}
\title{gemini_symbols}
\usage{
gemini_symbols()
}
\value{
returns a vector containing all symbols available on Gemini
}
\description{
gemini_symbols
}
\examples{
\dontrun{
symbols <- gemini_symbols()}
}
