% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gemini_trades}
\alias{gemini_trades}
\title{gemini_trades}
\usage{
gemini_trades(secret, key)
}
\arguments{
\item{secret}{your secret key for Gemini}

\item{key}{your API key for Gemini}
}
\value{
returns a dataframe containing all of your historical trades.
}
\description{
gemini_trades
}
\examples{
\dontrun{
secret <- "..."
key <- "..."
df <- gemini_trades(secret, key)}
}
