% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_block_at_hash}
\alias{helium_block_at_hash}
\title{helium_block_at_hash}
\usage{
helium_block_at_hash(hash)
}
\arguments{
\item{hash}{block hash for the block to fetch}
}
\value{
returns block descriptor for the given block hash
}
\description{
helium_block_at_hash
}
\examples{
\dontrun{
hash <- 'o5o8cBL1Zvp1KmA1d8vYJ38wXXWVxf1HW2e1uZxVJYg'
block_at_hash <- helium_block_at_hash(hash)}
}
