% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_block_at_height}
\alias{helium_block_at_height}
\title{helium_block_at_height}
\usage{
helium_block_at_height(height)
}
\arguments{
\item{height}{the blockchain height}
}
\value{
returns a list containing descriptors for the block at specified height
}
\description{
helium_block_at_height
}
\examples{
\dontrun{
height <- '213787'
block <- helium_block_at_height(height)}
}
