% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_block_descriptions}
\alias{helium_block_descriptions}
\title{helium_block_descriptions}
\usage{
helium_block_descriptions(cursor = NULL)
}
\arguments{
\item{cursor}{used to retrieve more results. Default value is NULL.}
}
\value{
returns a list contaning block descriptions. a cursor field will be
returned if there are more results.
}
\description{
helium_block_descriptions
}
\examples{
\dontrun{
block_descriptions <- helium_block_descriptions()

cursor <- block_descriptions$cursor
block_descriptions <- helium_block_descriptions(cursor)}
}
