% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_block_stats}
\alias{helium_block_stats}
\title{helium_block_stats}
\usage{
helium_block_stats()
}
\value{
returns a list containing block production statistics
}
\description{
helium_block_stats
}
\examples{
\dontrun{
block_stats <- helium_block_stats()}
}
