% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_blockchain_stats}
\alias{helium_blockchain_stats}
\title{helium_blockchain_stats}
\usage{
helium_blockchain_stats()
}
\value{
returns a list containing statistics about the Helium blockchain
}
\description{
helium_blockchain_stats
}
\examples{
\dontrun{
stats <- helium_blockchain_stats()}
}
