% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_height}
\alias{helium_height}
\title{helium_height}
\usage{
helium_height()
}
\value{
returns the height of the Helium blockchain
}
\description{
helium_height
}
\examples{
\dontrun{
height <- helium_height()}
}
