% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_transactions_at_height}
\alias{helium_transactions_at_height}
\title{helium_transactions_at_height}
\usage{
helium_transactions_at_height(height, cursor = NULL)
}
\arguments{
\item{height}{the blockchain height}

\item{cursor}{used to retrieve more results. Default value is NULL.}
}
\value{
returns a list containing block production statistics
}
\description{
helium_transactions_at_height
}
\examples{
\dontrun{
height <- '213787'
block_stats <- helium_block_descriptions(height)

cursor <- block_stats$cursor
block_stats <- helium_block_descriptions(height, cursor)}
}
