% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huobi.R
\name{huobi_candles}
\alias{huobi_candles}
\title{huobi_candles}
\usage{
huobi_candles(period, size, symbol)
}
\arguments{
\item{period}{the period of each candle. The following are acceptable options:
"1min", "5min", "15min", "30min", "60min", "4hour", "1day", "1mon", "1week",
"1year"}

\item{size}{the number of datapoints to return. This should fall between 1
and 2000.}

\item{symbol}{the trading symbol to query.}
}
\value{
returns a dataframe containing Huobi candle data
}
\description{
huobi_candles
}
\examples{
\dontrun{
period <- '1day'
size <- '200'
symbol <- 'btcusdt'
candles <- huobi_candles(period, size, symbol)}
}
