% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{kraken_ticker_info}
\alias{kraken_ticker_info}
\title{kraken_ticker_info}
\usage{
kraken_ticker_info(pair = NULL, timeout_seconds = 60)
}
\arguments{
\item{pair}{optionally provide one or more comma-separated asset pairs.}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a list containing ticker info for assets on Kraken. Refer to
Kraken for help interpreting response data:
https://docs.kraken.com/rest/#tag/Market-Data/operation/getTickerInformation
}
\description{
kraken_ticker_info
}
\examples{
kraken_ticker_info("ETHUSD", 4.5)
}
