% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kucoin.R
\name{kucoin_accounts}
\alias{kucoin_accounts}
\title{kucoin_accounts}
\usage{
kucoin_accounts(
  api_key,
  api_secret,
  passphrase,
  version = "2",
  timeout_seconds = 60
)
}
\arguments{
\item{api_key}{your Kucoin API key.}

\item{api_secret}{your Kucoin API secret.}

\item{passphrase}{the passphrase you created when you created your Kucoin API
key.}

\item{version}{your API key version. This can be retrieved from your Kucoin
API console. The default value is "2".}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe containing your Kucoin accounts and balances.
}
\description{
kucoin_accounts
}
\examples{
\dontrun{
api_key <- "..."
api_secret <- "..."
passphrase <- "..."
accounts <- kucoin_accounts(api_key, api_secret, passphrase)}
}
