% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kucoin.R
\name{kucoin_signature}
\alias{kucoin_signature}
\title{kucoin_signature}
\usage{
kucoin_signature(api_secret, time, method, path, body)
}
\arguments{
\item{api_secret}{your Kucoin API secret}

\item{time}{a timestamp string formatted the way Kucoin requires. This can be
created with the "kucoin_time" function.}

\item{method}{"GET" or "POST"}

\item{path}{the endpoint you are using to make an API call.}

\item{body}{needs to be a json string which matches url parameters. Use a
blank string if not applicable.}
}
\value{
returns a signature for use in you Kucoin API calls.
}
\description{
kucoin_signature
}
\examples{
\dontrun{
api_secret <- "..."
time <- kucoin_time()
method <- "GET"
path <- "/api/v1/sub/user"
body <- ""
sig <- kucoin_signature(api_secret, time, method, path, body)}
}
