% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kucoin.R
\name{kucoin_time}
\alias{kucoin_time}
\title{kucoin_time}
\usage{
kucoin_time()
}
\value{
returns a timestamp formatted in the way it is required in order to
make an API call to Kucoin.
}
\description{
kucoin_time
}
\examples{
kucoin_time()
}
