% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eden.R
\name{magic_eden_collection_stats}
\alias{magic_eden_collection_stats}
\title{magic_eden_collection_stats}
\usage{
magic_eden_collection_stats(symbol, timeout_seconds = 60)
}
\arguments{
\item{symbol}{the collection symbol you are requesting data for}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a list containing statistics about the specified collection.
}
\description{
magic_eden_collection_stats
}
\examples{
symbol <- "gothic_degens"
magic_eden_collection_stats(symbol, timeout_seconds = 4.5)
}
