% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eden.R
\name{magic_eden_token_listings}
\alias{magic_eden_token_listings}
\title{magic_eden_token_listings}
\usage{
magic_eden_token_listings(mint_address)
}
\arguments{
\item{mint_address}{the mint address of the token you wish to query}
}
\value{
returns a dataframe containing the token listings for the specified
mint address.
}
\description{
magic_eden_token_listings
}
\examples{
mint_address <- "Hd6sxFEEQQA5aURaWaDesi23AkM19bBkKave1hyWvnfS"
magic_eden_token_listings(mint_address)
}
