% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eden.R
\name{magic_eden_tokens_owned}
\alias{magic_eden_tokens_owned}
\title{magic_eden_tokens_owned}
\usage{
magic_eden_tokens_owned(
  wallet,
  offset = NULL,
  limit = NULL,
  list_status = NULL,
  timeout_seconds = 60
)
}
\arguments{
\item{wallet}{the address of the wallet you are trying to query}

\item{offset}{optionally provide a numeric value to specify number of
results to skip.}

\item{limit}{optionally provide a numeric limit to specify maximum number of
results.}

\item{list_status}{either "listed", "unlisted" or "both". The default is
"both".}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe containing all tokens owned by specified wallet.
}
\description{
magic_eden_tokens_owned
}
\examples{
wallet <- "72tXz6jhGVPFE8ZfAQocJPJU3HgxsdrRqKZoUdWUhs7o"
magic_eden_tokens_owned(wallet, timeout_seconds = 4.5)
}
