% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eden.R
\name{magic_eden_transactions}
\alias{magic_eden_transactions}
\title{magic_eden_transactions}
\usage{
magic_eden_transactions(wallet)
}
\arguments{
\item{wallet}{the address of the wallet you are trying to query}
}
\value{
returns a dataframe containing all transactions for the specified wallet.
}
\description{
magic_eden_transactions
}
\examples{
\dontrun{
wallet <- "72tXz6jhGVPFE8ZfAQocJPJU3HgxsdrRqKZoUdWUhs7o"
transactions <- magic_eden_transactions(wallet)}
}
