% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifty_gateway.R
\name{nifty_gateway_creators}
\alias{nifty_gateway_creators}
\title{nifty_gateway_creators}
\usage{
nifty_gateway_creators(username, limit = NULL, offset = NULL)
}
\arguments{
\item{username}{the username you wish to query}

\item{limit}{optionally provide the maximum number of results to return. This
is a numeric parameter.}

\item{offset}{optionally specify how many results to skip. This is a numeric
parameter.}
}
\value{
returns a dataframe containing ownership information for all nifties
created by the given creator and currently stored on Nifty Gateway
}
\description{
nifty_gateway_creators
}
\examples{
\dontrun{
creators <- nifty_gateway_creators('beeple')}
}
