% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okcoin.R
\name{okcoin_orders}
\alias{okcoin_orders}
\title{okcoin_orders}
\usage{
okcoin_orders(
  secret,
  key,
  passphrase,
  instrument_id,
  state,
  timeout_seconds = 60
)
}
\arguments{
\item{secret}{your secret key for Okcoin}

\item{key}{your API key for Okcoin}

\item{passphrase}{the passphrase which you created when generating your
Okcoin API key}

\item{instrument_id}{the trading pair symbol}

\item{state}{Order Status: -1: Canceled, 0: Open, 1: Partially Filled,
2: Fully Filled, 3: Submitting, 4: Canceling, 6: Incomplete (open + partially
filled), 7: Complete (canceled + fully filled)}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe containing your orders from the most recent 3
months
}
\description{
okcoin_orders
}
\examples{
\dontrun{
secret <- "..."
key <- "..."
passphrase <- "..."
instrument_id <- "BTC-USDT"
state <- '2'
orders <- okcoin_orders(secret, key, passphrase, instrument_id, state)}
}
