% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okcoin.R
\name{okcoin_signature}
\alias{okcoin_signature}
\title{okcoin_signature}
\usage{
okcoin_signature(path, secret, formatted_time, method)
}
\arguments{
\item{path}{the API endpoint}

\item{secret}{your Okcoin secret key}

\item{formatted_time}{a string containing the currrent timestamp in ISO 8601
format}

\item{method}{'POST' or 'GET'}
}
\value{
returns a base64 encoded SHA256 signature for signing Okcoin API
calls
}
\description{
okcoin_signature
}
\examples{
\dontrun{
path <- "..."
secret <- "..."
formatted_time <- okcoin_time()
method <- "GET"
signature <- okcoin_signature(path, secret, formatted_time, method)}
}
