% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okcoin.R
\name{okcoin_spot_account_info}
\alias{okcoin_spot_account_info}
\title{okcoin_spot_account_info}
\usage{
okcoin_spot_account_info(secret, key, passphrase)
}
\arguments{
\item{secret}{your secret key for Okcoin}

\item{key}{your API key for Okcoin}

\item{passphrase}{the passphrase which you created when generating your Okcoin API key}
}
\value{
returns a dataframe containing your spot account balances
}
\description{
okcoin_spot_account_info
}
\examples{
\dontrun{
secret <- "..."
key <- "..."
passphrase <- "..."
balances <- okcoin_spot_account_info(secret, key, passphrase)}
}
