% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okcoin.R
\name{okcoin_trading_pairs}
\alias{okcoin_trading_pairs}
\title{okcoin_trading_pairs}
\usage{
okcoin_trading_pairs()
}
\value{
returns a dataframe containing information about all trading pairs on Okcoin
}
\description{
okcoin_trading_pairs
}
\examples{
\dontrun{
trading_pairs <- okcoin_trading_pairs()}
}
