% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paxos.R
\name{paxos_bearer_token}
\alias{paxos_bearer_token}
\title{paxos_bearer_token}
\usage{
paxos_bearer_token(client_id, client_secret, scope, timeout_seconds = 60)
}
\arguments{
\item{client_id}{the client id you generated when you created your API key}

\item{client_secret}{the client secret you generated when you created your
API key}

\item{scope}{the scope needed for your specific API call}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns your Paxos bearer token
}
\description{
paxos_bearer_token
}
\examples{
\dontrun{
client_id <- "..."
client_secret <- "..."
scope <- 'funding:read_address'
token <- paxos_bearer_token(client_id, client_secret, scope)}
}
