% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paxos.R
\name{paxos_list_profiles}
\alias{paxos_list_profiles}
\title{paxos_list_profiles}
\usage{
paxos_list_profiles(client_id, client_secret, timeout_seconds = 60)
}
\arguments{
\item{client_id}{the client id you generated when you created your API key}

\item{client_secret}{the client secret you generated when you created your
API key}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe containing all user profiles
}
\description{
paxos_list_profiles
}
\examples{
\dontrun{
client_id <- "..."
client_secret <- "..."
profiles <- paxos_list_profiles(client_id, client_secret)}
}
