% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_api_call}
\alias{solana_api_call}
\title{solana_api_call}
\usage{
solana_api_call(url, request_body)
}
\arguments{
\item{url}{the RPC url for your API call}

\item{request_body}{the request body for your API call}
}
\value{
returns data from your solana API call
}
\description{
solana_api_call
}
\examples{
\dontrun{
url <- "https://api.devnet.solana.com"
data <- solana_api_call(url)}
}
