% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_assemble_key_pair}
\alias{solana_assemble_key_pair}
\title{solana_assemble_key_pair}
\usage{
solana_assemble_key_pair(key, pair)
}
\arguments{
\item{key}{the key for your key pair}

\item{pair}{the pair for your key pair}
}
\value{
Returns your key pair if it exists or a blank string if it doesn't
exist
}
\description{
solana_assemble_key_pair
}
\examples{
limit <- NULL
limit <- solana_assemble_key_pair('limit', limit)
}
