% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_assemble_list}
\alias{solana_assemble_list}
\title{solana_assemble_list}
\usage{
solana_assemble_list(character_vector)
}
\arguments{
\item{character_vector}{the character vector used to create the config object}
}
\value{
Returns your config object
}
\description{
solana_assemble_list
}
\examples{
limit <- solana_assemble_key_pair('limit', NULL)
character_vector <- c(limit)
config_object <- solana_assemble_list(character_vector)
}
