% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_assemble_request_body}
\alias{solana_assemble_request_body}
\title{solana_assemble_request_body}
\usage{
solana_assemble_request_body(jsonrpc, id, method, params)
}
\arguments{
\item{jsonrpc}{the jsonrpc for your request body}

\item{id}{the id for your request body}

\item{method}{the method for your request body}

\item{params}{the parameters for your request body}
}
\value{
Returns the request body for your solana API call
}
\description{
solana_assemble_request_body
}
\examples{
\dontrun{
limit <- solana_assemble_key_pair('limit', NULL)
character_vector <- c(limit)
config_object <- solana_assemble_list(character_vector)
address <- "Vote111111111111111111111111111111111111111"
params <- paste('["', address, '", {', config_object, '}]', sep = '')
solana_assemble_request_body('"2.0"', 'null', '"getSignaturesForAddress"', params)}
}
