% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_block}
\alias{solana_get_block}
\title{solana_get_block}
\usage{
solana_get_block(url, slot, timeout_seconds = 60)
}
\arguments{
\item{url}{the RPC url for your API call}

\item{slot}{slot number, as u64 integer}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
Returns identity and transaction information about a confirmed block
in the ledger.
}
\description{
solana_get_block
}
\examples{
\dontrun{
url <- "https://api.devnet.solana.com"
slot <- solana_get_slot(url)
data <- solana_get_block(url, slot)}
}
