% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_genesis_hash}
\alias{solana_get_genesis_hash}
\title{solana_get_genesis_hash}
\usage{
solana_get_genesis_hash(url, timeout_seconds = 60)
}
\arguments{
\item{url}{the RPC url for your API call}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
Returns the genesis hash
}
\description{
solana_get_genesis_hash
}
\examples{
\donttest{
url <- "https://api.devnet.solana.com"
data <- solana_get_genesis_hash(url)}
}
