% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_health}
\alias{solana_get_health}
\title{solana_get_health}
\usage{
solana_get_health(url)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the current health of the node.
}
\description{
solana_get_health
}
\examples{
\dontrun{
url <- "https://api.devnet.solana.com"
data <- solana_get_health(url)}
}
