% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_inflation_rate}
\alias{solana_get_inflation_rate}
\title{solana_get_inflation_rate}
\usage{
solana_get_inflation_rate(url)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the specific inflation values for the current epoch
}
\description{
solana_get_inflation_rate
}
\examples{
url <- "https://api.devnet.solana.com"
data <- solana_get_inflation_rate(url)
}
