% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_slot}
\alias{solana_get_slot}
\title{solana_get_slot}
\usage{
solana_get_slot(url)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the slot that has reached the given or default commitment
level.
}
\description{
solana_get_slot
}
\examples{
\dontrun{
url <- "https://api.devnet.solana.com"
data <- solana_get_slot(url)}
}
