% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_us.R
\name{binance_us_ping}
\alias{binance_us_ping}
\title{binance_us_ping}
\usage{
binance_us_ping()
}
\value{
returns a response from the Binance.US API server
}
\description{
binance_us_ping
}
\examples{
binance_us_ping()
}
