% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_us.R
\name{binance_us_recent_trades}
\alias{binance_us_recent_trades}
\title{binance_us_recent_trades}
\usage{
binance_us_recent_trades(symbol, limit)
}
\arguments{
\item{symbol}{the trading pair for which you wish to retrieve data.}

\item{limit}{the number of results to return. The maximum is 1,000.}
}
\value{
returns a dataframe containing the most recent trades executed for the designated currency pair on Binance US
}
\description{
binance_us_recent_trades
}
\examples{
\dontrun{
symbol <- 'LTCBTC'
limit <- '1000'
recent_trades <- binance_us_recent_trades(symbol, limit)}
}
