% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_us.R
\name{binance_us_server_time}
\alias{binance_us_server_time}
\title{binance_us_server_time}
\usage{
binance_us_server_time(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns the Binance.US server time
}
\description{
binance_us_server_time
}
\examples{
binance_us_server_time(4.5)
}
