% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_us.R
\name{binance_us_system_status}
\alias{binance_us_system_status}
\title{binance_us_system_status}
\usage{
binance_us_system_status(key, secret, timeout_seconds = 60)
}
\arguments{
\item{key}{your Binance.US API key}

\item{secret}{your Binance.US secret key}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns the status of the Binance.US API. The response will either be
a "0" for normal or a "1" for system maintenance.
}
\description{
binance_us_system_status
}
\examples{
\dontrun{
key <- "..."
secret <- "..."
system_status <- binance_us_system_status(key, secret)}
}
