% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain_dot_com.R
\name{blockchain_dot_com_ticker_symbol}
\alias{blockchain_dot_com_ticker_symbol}
\title{blockchain_dot_com_ticker_symbol}
\usage{
blockchain_dot_com_ticker_symbol(symbol)
}
\value{
returns a list with price and volume data for specified symbol
}
\description{
blockchain_dot_com_ticker_symbol
}
\examples{
\dontrun{
symbol <- 'BTC-USD'
ticker_data <- blockchain_dot_com_ticker_symbol(symbol)}
}
