% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cex_io.R
\name{cex_io_ohlcv}
\alias{cex_io_ohlcv}
\title{cex_io_ohlcv}
\usage{
cex_io_ohlcv(date, symbol_1, symbol_2)
}
\arguments{
\item{date}{the date for which to retrieve data}

\item{symbol_1}{the first currency in your pair}

\item{symbol_2}{the second currency in your pair}
}
\value{
returns a list containing open, high, low, close, volume data for the past minute, hour, and day
}
\description{
cex_io_ohlcv
}
\examples{
\dontrun{
date <- '20220927'
symbol_1 <- 'btc'
symbol_2 <- 'usd'
ohlcv <- cex_io_ohlcv(date, symbol_1, symbol_2)}
}
