% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase.R
\name{coinbase_api_call}
\alias{coinbase_api_call}
\title{coinbase_api_call}
\usage{
coinbase_api_call(
  api_key,
  api_secret,
  method,
  path,
  body,
  query = NULL,
  timeout_seconds = 60
)
}
\arguments{
\item{api_key}{your Coinbase API key}

\item{api_secret}{your Coinbase API secret}

\item{method}{"GET" or "POST"}

\item{path}{the path of your API call}

\item{body}{the body of your API call}

\item{query}{the query for your coinbase API call as a list}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns the response from your Coinbase API call
}
\description{
coinbase_api_call
}
\examples{
\dontrun{
path <- "/api/v3/brokerage/accounts"
method <- "GET"
api_key <- "..."
api_secret <- "..."
body <- ""
data <- coinbase_api_call(api_key, api_secret, method, path, body)}
}
