% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko.R
\name{coingecko_categories}
\alias{coingecko_categories}
\title{coingecko_categories}
\usage{
coingecko_categories()
}
\value{
returns a dataframe of all categories on CoinGecko.
}
\description{
coingecko_categories
}
\examples{
\dontrun{
categories <- coingecko_categories()}
}
