% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko.R
\name{coingecko_global_data}
\alias{coingecko_global_data}
\title{coingecko_global_data}
\usage{
coingecko_global_data()
}
\value{
returns a list containing high-level statistics about the
cryptocurrency ecosystem.
}
\description{
coingecko_global_data
}
\examples{
\dontrun{
global_data <- coingecko_global_data()}
}
