% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko.R
\name{coingecko_ping}
\alias{coingecko_ping}
\title{coingecko_ping}
\usage{
coingecko_ping(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns the Coingecko API server status
}
\description{
coingecko_ping
}
\examples{
coingecko_ping(4.5)
}
