% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko.R
\name{coingecko_price_history}
\alias{coingecko_price_history}
\title{coingecko_price_history}
\usage{
coingecko_price_history(id, date, localization = "false", timeout_seconds = 60)
}
\arguments{
\item{id}{The asset id you wish to query. IDs can be retrieved with the
coingecko_coins function.}

\item{date}{the date you wish to query formatted as "dd-mm-yyyy"}

\item{localization}{"true" or "false" to include/exclude localized languages
in the response. The default value is "false".}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a list containing data about asset pricing.
}
\description{
coingecko_price_history
}
\examples{
price <- coingecko_price_history("bitcoin", "30-12-2017", timeout_seconds = 4.5)
price$market_data$current_price$usd
}
