% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko.R
\name{coingecko_vs_currencies}
\alias{coingecko_vs_currencies}
\title{coingecko_vs_currencies}
\usage{
coingecko_vs_currencies()
}
\value{
returns a character vector containing all supported currencies on
Coingecko.
}
\description{
coingecko_vs_currencies
}
\examples{
coingecko_vs_currencies()
}
