% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinlist.R
\name{coinlist_api_call}
\alias{coinlist_api_call}
\title{coinlist_api_call}
\usage{
coinlist_api_call(
  api_key,
  api_secret,
  method,
  path,
  body,
  timeout_seconds = 60
)
}
\arguments{
\item{api_key}{your Coinlist API key}

\item{api_secret}{your Coinlist API secret}

\item{method}{"GET" or "POST"}

\item{path}{the path of your API call}

\item{body}{the body of your API call}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns the response from your Coinlist API call
}
\description{
coinlist_api_call
}
\examples{
\dontrun{
path <- "/v1/accounts"
method <- "GET"
api_key <- "..."
api_secret <- "..."
body <- ""
data <- coinlist_api_call(api_key, api_secret, method, path, body)}
}
