% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinlist.R
\name{coinlist_symbols}
\alias{coinlist_symbols}
\title{coinlist_symbols}
\usage{
coinlist_symbols(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe with information about symbols available on
Coinlist Pro
}
\description{
coinlist_symbols
}
\examples{
coinlist_symbols(4.5)
}
