% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_dot_com.R
\name{crypto_dot_com_get_book}
\alias{crypto_dot_com_get_book}
\title{crypto_dot_com_get_book}
\usage{
crypto_dot_com_get_book(instrument, depth = 50)
}
\arguments{
\item{instrument}{the instrument name which you want to query}

\item{depth}{the depth of the order book to retrieve. The maximum and default
value is 50.}
}
\value{
returns a list containing the order book for your specified
instrument.
}
\description{
crypto_dot_com_get_book
}
\examples{
\dontrun{
book <- crypto_dot_com_get_book("BTC_USDT")}
}
