% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_dot_com.R
\name{crypto_dot_com_instruments}
\alias{crypto_dot_com_instruments}
\title{crypto_dot_com_instruments}
\usage{
crypto_dot_com_instruments(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe with information about instruments available on
Crypto.com
}
\description{
crypto_dot_com_instruments
}
\examples{
crypto_dot_com_instruments(4.5)
}
