% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gemini_symbols}
\alias{gemini_symbols}
\title{gemini_symbols}
\usage{
gemini_symbols(timeout_seconds = 60)
}
\arguments{
\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a vector containing all symbols available on Gemini
}
\description{
gemini_symbols
}
\examples{
gemini_symbols(4.5)
}
