% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huobi.R
\name{huobi_candles}
\alias{huobi_candles}
\title{huobi_candles}
\usage{
huobi_candles(period, size, symbol, timeout_seconds = 60)
}
\arguments{
\item{period}{the period of each candle. The following are acceptable options:
"1min", "5min", "15min", "30min", "60min", "4hour", "1day", "1mon", "1week",
"1year"}

\item{size}{the number of datapoints to return. This should fall between 1
and 2000.}

\item{symbol}{the trading symbol to query.}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe containing Huobi candle data
}
\description{
huobi_candles
}
\examples{
period <- '1day'
size <- '200'
symbol <- 'btcusdt'
huobi_candles(period, size, symbol, 4.5)
}
