% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{kraken_asset_info}
\alias{kraken_asset_info}
\title{kraken_asset_info}
\usage{
kraken_asset_info(asset = NULL, aclass = NULL)
}
\arguments{
\item{asset}{optionally provide one or more comma-separated ticker symbols.}

\item{aclass}{optionally provide asset categories to filter by.}
}
\value{
returns a list containing asset information
}
\description{
kraken_asset_info
}
\examples{
all_asset_info <- kraken_asset_info()
eth_btc_info <- kraken_asset_info("ETH,BTC")
currency_info <- kraken_asset_info(aclass = "currency")
}
