% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{kraken_asset_pairs}
\alias{kraken_asset_pairs}
\title{kraken_asset_pairs}
\usage{
kraken_asset_pairs(pair = NULL, info = NULL)
}
\arguments{
\item{pair}{optionally provide one or more comma-separated asset pairs to query.}

\item{info}{optionally select the information to return. You can choose from:
"info" (all info), "leverage" (leverage info), "fees" (fee schedule), or
"margin" (margin info).}
}
\value{
returns a list containing information on Kraken asset pairs.
}
\description{
kraken_asset_pairs
}
\examples{
\dontrun{
asset_pairs <- kraken_asset_pairs()}
}
