% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{kraken_server_status}
\alias{kraken_server_status}
\title{kraken_server_status}
\usage{
kraken_server_status()
}
\value{
returns a list with Kraken's server status along with a timestamp
}
\description{
kraken_server_status
}
\examples{
kraken_server_status()
}
