% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{kraken_server_time}
\alias{kraken_server_time}
\title{kraken_server_time}
\usage{
kraken_server_time()
}
\value{
returns a list with Kraken's server time in unix and rfc1123 formats
}
\description{
kraken_server_time
}
\examples{
\dontrun{
time <- kraken_server_time()}
}
