% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kucoin.R
\name{kucoin_symbols_list}
\alias{kucoin_symbols_list}
\title{kucoin_symbols_list}
\usage{
kucoin_symbols_list(market = NULL)
}
\arguments{
\item{market}{optionally provide a market to filter on. This function will
return all markets by default.}
}
\value{
returns a dataframe containing information about trading symbols
}
\description{
kucoin_symbols_list
}
\examples{
\dontrun{
symbols <- kucoin_symbols_list('btc')}
}
