% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eden.R
\name{magic_eden_collection_stats}
\alias{magic_eden_collection_stats}
\title{magic_eden_collection_stats}
\usage{
magic_eden_collection_stats(symbol)
}
\arguments{
\item{symbol}{the collection symbol you are requesting data for}
}
\value{
returns a list containing statistics about the specified collection.
}
\description{
magic_eden_collection_stats
}
\examples{
symbol <- "gothic_degens"
magic_eden_collection_stats(symbol)
}
