% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eden.R
\name{magic_eden_transactions}
\alias{magic_eden_transactions}
\title{magic_eden_transactions}
\usage{
magic_eden_transactions(
  wallet,
  offset = NULL,
  limit = NULL,
  timeout_seconds = 60
)
}
\arguments{
\item{wallet}{the address of the wallet you are trying to query}

\item{offset}{optionally provide a numeric value to specify number of
transactions to skip.}

\item{limit}{optionally provide a numeric limit to specify maximum number of
results.}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns a dataframe containing all transactions for the specified
wallet.
}
\description{
magic_eden_transactions
}
\examples{
wallet <- "72tXz6jhGVPFE8ZfAQocJPJU3HgxsdrRqKZoUdWUhs7o"
magic_eden_transactions(wallet, timeout_seconds = 4.5)
}
