% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_api_call}
\alias{solana_api_call}
\title{solana_api_call}
\usage{
solana_api_call(url, request_body, timeout_seconds = 60)
}
\arguments{
\item{url}{the RPC url for your API call}

\item{request_body}{the request body for your API call}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
returns data from your Solana API call
}
\description{
solana_api_call
}
\examples{
\donttest{
url <- "https://api.devnet.solana.com"
request_body <-
  solana_assemble_request_body('"2.0"', 'null', '"getBlockHeight"', NULL)
data <- solana_api_call(url, request_body)}
}
