% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_assemble_list}
\alias{solana_assemble_list}
\title{solana_assemble_list}
\usage{
solana_assemble_list(character_vector)
}
\arguments{
\item{character_vector}{the character vector used to create the config object}
}
\value{
Returns your config object
}
\description{
solana_assemble_list
}
\examples{
jsonrpc <- solana_assemble_key_pair('jsonrpc', jsonrpc)
id <- solana_assemble_key_pair('id', id)
method <- solana_assemble_key_pair('method', method)
params <- solana_assemble_key_pair('params', params)
character_vector <- c(jsonrpc, id, method, params)
body <- solana_assemble_list(character_vector)
request_body <- paste('{', body, '}', sep = '')
}
