% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_assemble_request_body}
\alias{solana_assemble_request_body}
\title{solana_assemble_request_body}
\usage{
solana_assemble_request_body(jsonrpc, id, method, params)
}
\arguments{
\item{jsonrpc}{the jsonrpc for your request body}

\item{id}{the id for your request body}

\item{method}{the method for your request body}

\item{params}{the parameters for your request body}
}
\value{
Returns the request body for your solana API call
}
\description{
solana_assemble_request_body
}
\examples{
\dontrun{
limit <- solana_assemble_key_pair('limit', limit)
params <- paste('["',address,'", {',limit,'}]', sep = '')
request_body <- solana_assemble_request_body('2.0', 'null', 'getSignaturesForAddress', params)}
}
