% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_genesis_hash}
\alias{solana_get_genesis_hash}
\title{solana_get_genesis_hash}
\usage{
solana_get_genesis_hash(url)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the genesis hash
}
\description{
solana_get_genesis_hash
}
\examples{
url <- "https://api.devnet.solana.com"
data <- solana_get_genesis_hash(url)
}
