% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_identity}
\alias{solana_get_identity}
\title{solana_get_identity}
\usage{
solana_get_identity(url)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the identity pubkey for the current node
}
\description{
solana_get_identity
}
\examples{
\dontrun{
url <- "https://api.devnet.solana.com"
data <- solana_get_identity(url)}
}
