% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_inflation_rate}
\alias{solana_get_inflation_rate}
\title{solana_get_inflation_rate}
\usage{
solana_get_inflation_rate(url, timeout_seconds = 60)
}
\arguments{
\item{url}{the RPC url for your API call}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
Returns the specific inflation values for the current epoch
}
\description{
solana_get_inflation_rate
}
\examples{
\dontrun{
url <- "https://api.devnet.solana.com"
data <- solana_get_inflation_rate(url)}
}
