% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_recent_prioritization_fees}
\alias{solana_get_recent_prioritization_fees}
\title{solana_get_recent_prioritization_fees}
\usage{
solana_get_recent_prioritization_fees(url)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns a list of prioritization fees from recent blocks.
}
\description{
solana_get_recent_prioritization_fees
}
\examples{
url <- "https://api.devnet.solana.com"
data <- solana_get_recent_prioritization_fees(url)
}
