% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_slot}
\alias{solana_get_slot}
\title{solana_get_slot}
\usage{
solana_get_slot(url, timeout_seconds = 60)
}
\arguments{
\item{url}{the RPC url for your API call}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
Returns the slot that has reached the given or default commitment
level.
}
\description{
solana_get_slot
}
\examples{
\donttest{
url <- "https://api.devnet.solana.com"
data <- solana_get_slot(url)}
}
