% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{solana_get_supply}
\alias{solana_get_supply}
\title{solana_get_supply}
\usage{
solana_get_supply(url, timeout_seconds = 60)
}
\arguments{
\item{url}{the RPC url for your API call}

\item{timeout_seconds}{seconds until the query times out. Default is 60.}
}
\value{
Returns information about the current supply.
}
\description{
solana_get_supply
}
\examples{
\dontrun{
url <- "https://api.devnet.solana.com"
data <- solana_get_supply(url)}
}
